правки в интерфейсе и коде алгоритмов
фикс пользовательского ввода полинома, теперь можно вводить 16-ричные числа
фикс ошибок вертикального и горизонтального контроля по паритету
добавлено окно "О программе"
небольшой cleanup

+++ Lab3Seti/CRCRefactoring.cs
@@ -10,9 +10,9 @@ namespace Lab3Seti
     //тоже самое (почти) что и в классе CRCFromC но более понятным языком
     public static class CRCRefactoring
     {
-        public static void CRC32(char[] arrOrig, out uint ctrlSum, int degreePolynom, uint polymome)
-        { 
-            uint _register = 0x0;
+        public static void CRC32(char[] arrOrig, out uint ctrlSum, int degreePolynom, ulong polymome)
+        {
+            ulong _register = 0x0;
             uint _bitMask = (uint)(Math.Pow(2, degreePolynom) - 1); // маска для удаления лишнего байта, кол-во единиц = степени полинома
 
             char[] message = new char[arrOrig.Length + (degreePolynom / 8)];
@@ -27,23 +27,23 @@ namespace Lab3Seti
                 int bitPositionFinal = 8;
                 while (bitPosition != bitPositionFinal)
                 {
-                    uint BitIn = GetBit(bitPositionFinal, bitPosition, message[i]);
-                    uint BitOut = GetBit(degreePolynom, 0, _register);
+                    ulong BitIn = GetBit(bitPositionFinal, bitPosition, message[i]);
+                    ulong BitOut = GetBit(degreePolynom, 0, _register);
                     _register = RegistorPushAndXOR(BitIn, BitOut, _register, polymome, _bitMask);
                     bitPosition++;
                     
                 }
             }
-            ctrlSum = _register & _bitMask;
+            ctrlSum = (uint)_register & _bitMask;
         }
 
-        private static uint GetBit(int countBit, int position, uint word)
+        private static ulong GetBit(int countBit, int position, ulong word)
         {
             byte bitMask = 1;
             return word >> (countBit - position - 1) & bitMask;
         }
 
-        private static uint RegistorPushAndXOR(uint inBit, uint outBit, uint register, uint polinome, uint bitMask)
+        private static ulong RegistorPushAndXOR(ulong inBit, ulong outBit, ulong register, ulong polinome, uint bitMask)
         {
             register = (register << 1) & bitMask | inBit;
             if (outBit == 1)
